requirements.txt
numpy-stl
streamlit
plotly
numpy
trimesh
rtree
uuid
io
time